Finalize update structure. Update make files
