+++ miniapps/lpq-jacobi/CMakeLists.txt
@@ -13,7 +13,6 @@ if(MFEM_USE_MPI)
   set(MESH_FILES_FROM_MESHING
     icf.mesh
     cube.mesh
-    # jagged.mesh
     # square01.mesh
     # square01-tri.mesh
     # stretched2D.mesh
@@ -48,9 +47,8 @@ if(MFEM_USE_MPI)
     list(APPEND SRC_MESH_FILES ${CMAKE_CURRENT_SOURCE_DIR}/../../data/${MESH_FILE})
   endforeach()
   
-  file(MAKE_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}/meshes)
-  
   add_custom_command(OUTPUT data_is_copied
+    COMMAND ${CMAKE_COMMAND} -E make_directory meshes
     COMMAND ${CMAKE_COMMAND} -E copy_if_different ${SRC_MESH_FILES} ./meshes
     COMMAND ${CMAKE_COMMAND} -E touch data_is_copied
     COMMENT "Copying meshing miniapps data files ...")
@@ -59,45 +57,16 @@ if(MFEM_USE_MPI)
   add_mfem_miniapp(lpq-jacobi
     MAIN lpq-jacobi.cpp
     ${MFEM_MINIAPPS_COMMON_HEADERS}
-    EXTRA_HEADERS lpq-jacobi.hpp
+    EXTRA_SOURCES lpq-common.cpp
+    EXTRA_HEADERS lpq-common.hpp
     LIBRARIES mfem-common mfem)
   add_dependencies(lpq-jacobi copy_meshes)
 
-  add_mfem_miniapp(ml-lpq-jacobi
-    MAIN ml-lpq-jacobi.cpp
+  add_mfem_miniapp(mg-lpq-jacobi
+    MAIN mg-lpq-jacobi.cpp
     ${MFEM_MINIAPPS_COMMON_HEADERS}
-    EXTRA_HEADERS lpq-jacobi.hpp
+    EXTRA_SOURCES lpq-common.cpp
+    EXTRA_HEADERS lpq-common.hpp
     LIBRARIES mfem-common mfem)
-  add_dependencies(ml-lpq-jacobi copy_meshes)
+  add_dependencies(mg-lpq-jacobi copy_meshes)
 endif()
-
-
-# TODO(Gabriel): Add parallel code and tests.
-# # Parallel apps.
-# if (MFEM_USE_MPI)
-#   add_mfem_miniapp(plpq-jacobi
-#     MAIN plpq-jacobi.cpp
-#     ${MFEM_MINIAPPS_COMMON_HEADERS}
-#     LIBRARIES mfem-common mfem)
-#   add_dependencies(plpq-jacobi copy_meshes)
-# 
-#   # Add parallel tests.
-#   if (MFEM_ENABLE_TESTING)
-#     set(PARALLEL_TESTS
-#       plpq-jacobi
-#     )
-#     # TODO(Gabriel): Check this
-#     # Meshing miniapps that return MFEM_SKIP_RETURN_VALUE in some cases:
-#     set(SKIP_TESTS)
-# 
-#     foreach(test ${PARALLEL_TESTS})
-#       if (test IN_LIST SKIP_TESTS)
-#         continue()
-#       endif()
-#       add_test(NAME ${test}_np=${MFEM_MPI_NP}
-#         COMMAND ${MPIEXEC} ${MPIEXEC_NUMPROC_FLAG} ${MFEM_MPI_NP}
-#         ${MPIEXEC_PREFLAGS} $<TARGET_FILE:${test}> -no-vis
-#         ${MPIEXEC_POSTFLAGS})
-#     endforeach()
-#   endif()
-# endif()
